ener-vault/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project]
name = "ener-vault"
version = "0.1.0"
description = "Energy meter measurements API"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.14.0",
    "fastapi>=0.115.0",
    "psycopg[binary]>=3.2.0",
    "pydantic-settings>=2.6.0",
    "pygination>=0.0.13",
    "sqlalchemy>=2.0.36",
    "uvicorn[standard]>=0.32.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.0",
    "pytest>=8.3.0",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py313"
